renaming App to dashboard as App -> app made git and hot loader completely insane.
